% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisbinom.R
\name{poisson.binomial.pdf}
\alias{poisson.binomial.pdf}
\title{Poisson Binomial PDF}
\usage{
poisson.binomial.pdf(p)
}
\arguments{
\item{p}{A numeric vector of binomial probabilities.}
}
\value{
A numeric matrix of outcomes and probabilities.
}
\description{
Poisson Binomial PDF
}
