% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pme.R
\name{pme}
\alias{pme}
\title{Calculate Point of Minimum Error}
\usage{
pme(x, p)
}
\arguments{
\item{x}{A matrix or data frame of binary responses to infrequency items.}

\item{p}{A vector of probabilities of random responses being flagged by
infrequency items.}
}
\value{
A list of class "pme" containing diagnostics and the cutoff point.
}
\description{
Calculate Point of Minimum Error
}
